refactor: split into config, app, state and lib
in order to untangle these and support headless simulation

ref https://github.com/mlevison/high-performance-teams-game/issues/103
