+++ src/app/components/CapStoryChart.module.css
