+++ src/app/components/EffectValue.module.css
