+++ src/app/components/Rules.tsx
