+++ src/app/index.tsx
@@ -1,0 +1,3 @@
+import './global.css';
+
+export { default } from './App';
