+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './global.css';
-import App from './App';
+import App from './app';
 import { getInitialState } from './lib';
 import { config } from './config';
