+++ src/state/gameActions/index.ts
@@ -1,12 +1,8 @@
-import { GameAction as GameActionT } from './types';
-import { GameActionWithStatus as GameActionWithStatusT } from './getAvailableGameActions';
+export type { GameAction, GameActionWithImage } from './types';
+export type { GameActionWithStatus } from './getAvailableGameActions';
 export {
   getAvailableGameActions,
   UNIQUE as UNIQUE_ACTION,
 } from './getAvailableGameActions';
 export { getEffects } from './getEffects';
 export { getCost } from './getCost';
-export { isGameActionWithIcon, isGameActionWithImage } from './types';
-
-export type GameAction = GameActionT;
-export type GameActionWithStatus = GameActionWithStatusT;
