+++ src/state/index.ts
@@ -3,17 +3,16 @@ import type { Action } from './game';
 export type { ClosedGameRound as ClosedRound, AppRound } from './round';
 export type { GameState, GameActionAction, GameConfig } from './game';
 export type { AppState } from './useAppState';
-export type { GameActionWithStatus, GameAction } from './gameActions';
+export type {
+  GameActionWithStatus,
+  GameAction,
+  GameActionWithImage,
+} from './gameActions';
 export type { RoundDescription } from './rounds';
 export type { GremlinList } from './gremlins';
 export type { GameEffect, VisibleEffect, BaseEffect, Effect } from './effects';
 export type GameDispatch = Dispatch<Action>;
 export { getCosts, closeRound } from './round';
-export {
-  getAvailableGameActions,
-  isGameActionWithIcon,
-  isGameActionWithImage,
-  UNIQUE_ACTION,
-} from './gameActions';
+export { getAvailableGameActions, UNIQUE_ACTION } from './gameActions';
 export { default as useAppState } from './useAppState';
 export { createGameReducer, INITIAL_STATE } from './game';
